# Conflicts:
#	src/main/java/ee/coffee/coffeeshop/controllers/OrderController.java
#	src/main/java/ee/coffee/coffeeshop/models/enums/OrderStatus.java
#	src/main/resources/templates/products.ftlh

+++ src/main/java/ee/coffee/coffeeshop/services/interfaces/ProductService.java
@@ -4,8 +4,6 @@ import ee.coffee.coffeeshop.entity.Product;
 import java.util.List;
 
 public interface ProductService {
-
     void addProduct(Product product);
-
-    void deleteProduct(Integer id);
+    void deleteProduct(Long id);
 }
